refactor: formatYenFromDecimal を (features)/_shared へ昇格
原価一覧でも同じ表示整形が必要になるため、共通売単価 _components から集約非依存の
_shared へ移動した。理由: DTO の「集約ごとに複製」規約（ADR-20260627-a5c）は集約固有の
型に対する規約で、集約非依存の表示用純関数は適用外。_shared/error-handler.ts の前例に従う。

移動前に現挙動を固定するユニットテスト（_shared/formatYen.test.ts）を新設。
共通売単価側の全 importer（5ファイル）の import を差し替え。
